Update improvement_algorithm.py
This code improves the original Tremaux algorithm by noting nodes having three or four open paths. Then, it prevents the robot from going around dead-end traps multiple times. If the node is a beginning of a trap, the code removes this node from the nodes list. Therefore, only safe nodes stay on the list.

+++ improvement_algorithm.py
@@ -55,8 +55,8 @@ elif node == 1:
         moves.append(left(i))
 
     leng = len(moves)
-    important_move = moves[(leng / 4) + 1]
-    other_move = moves[((leng / 4) + 1) + 1]
+    important_move = moves[int((leng / 4) + 1)]
+    other_move = moves[int(((leng / 4) + 1) + 1)]
 
 
     if (i == 0) and ("fwd" in moves and "right in moves" and "left" in moves and "back" in moves):
